move TX tests to own file
